requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
diskcache==5.6.1
duckduckgo-search==2.9.2
Flask==2.3.2
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
hugchat==0.0.5.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.154
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pyparsing==3.0.9
python-dotenv==1.0.0
PyYAML==6.0
requests==2.29.0
rsa==4.9
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.21.3
soupsieve==2.4.1
SQLAlchemy==2.0.12
sqlitedict==2.1.0
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.5.0
typing-inspect==0.8.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.3
wikipedia==1.4.0
yarl==1.9.2
zipp==3.15.0
